rfac: consolidate branch deployments
